requirements.txt
mysql-connector-python
requests
psycopg2-binary
